improving  for siesta

+++ emuhelper/siesta/static.py
@@ -20,37 +20,61 @@ class static_run:
         self.electrons = siesta_electrons()
         self.properties = siesta_properties()
         
-        self.electrons.xc["functional"] = "GGA"
-        self.electrons.xc["authors"] = "PBE"
-        self.electrons.dm["Tolerance"] = "1.d-6"
-        self.electrons.dm["MixingWight"] = 0.1
-        self.electrons.dm["NumberPulay"] = 8 # this can affect the convergence of scf
-        self.electrons.dm["AllowExtrapolation"] = "true"
-        self.electrons.dm["UseSaveDM"] = "false"
-        self.electrons.params["SolutionMethod"] = "diagon"
-        self.electrons.params["MeshCutoff"] = 300 #100
-        
-
-    def gen_input(self, directory="tmp-siesta-static", inpname="static.fdf"):
-        
-        if os.path.exists(directory):
-            shutil.rmtree(directory)
-        os.mkdir(directory)
-        
-        for element in self.system.xyz.specie_labels:
-            shutil.copyfile("%s.psf" % element, os.path.join(directory, "%s.psf" % element))
+        self.electrons.basic_setting()
 
                 
-        with open(os.path.join(directory, inpname), 'w') as fout:
-            self.system.to_fdf(fout)
-            self.electrons.to_fdf(fout)
-            self.properties.to_fdf(fout)
-    
-    def run(self, directory="tmp-siesta-static", inpname="static.fdf", output="static.out"):
-        # run the simulation
-        os.chdir(directory)
-        os.system("siesta < %s | tee %s" % (inpname, output))
-        os.chdir("../")
+
+    def scf(self, directory="tmp-siesta-static", inpname="static-scf.fdf", output="static-scf.out",
+            mpi="", runopt="gen", electrons={}, properties=0, kpoints_mp=[1, 1, 1]):
+        if runopt == "gen" or runopt == "genrun":
+            if os.path.exists(directory):
+                shutil.rmtree(directory)
+            os.mkdir(directory)
+        
+            for element in self.system.xyz.specie_labels:
+                shutil.copyfile("%s.psf" % element, os.path.join(directory, "%s.psf" % element))
+       
+            self.electrons.kpoints_mp = kpoints_mp
+            self.electrons.set_params(electrons)
+            # use self.properties.option to contorl the calculation of properties
+            self.properties.option = properties
+            with open(os.path.join(directory, inpname), 'w') as fout:
+                self.system.to_fdf(fout)
+                self.electrons.to_fdf(fout)
+                self.properties.to_fdf(fout)
+
+        if runopt == "run" or runopt == "genrun":
+            os.chdir(directory)
+            os.system("%s siesta < %s | tee %s" % (mpi, inpname, output))
+            os.chdir("../")
+
+    def scf_restart(self, directory="tmp-siesta-static", inpname="static-scf-restart.fdf", output="static-scf-restart.out",
+            mpi="", runopt="gen", electrons={}, properties=0, kpoints_mp=[1, 1, 1]):
+
+        # first check whether there is a previous scf running
+        if not os.path.exists(directory):
+            print("===================================================\n")
+            print("                 Warning !!!\n")
+            print("===================================================\n")
+            print("scf(restart) calculation:\n")
+            print("  directory of previous scf calculattion not found!\n")
+            sys.exit(1)
+        if runopt == "gen" or runopt == "genrun":
+       
+            self.electrons.dm["UseSaveDM"] = "true"
+            self.electrons.kpoints_mp = kpoints_mp
+            self.electrons.set_params(electrons)
+            # use self.properties.option to contorl the calculation of properties
+            self.properties.option = properties
+            with open(os.path.join(directory, inpname), 'w') as fout:
+                self.system.to_fdf(fout)
+                self.electrons.to_fdf(fout)
+                self.properties.to_fdf(fout)
+
+        if runopt == "run" or runopt == "genrun":
+            os.chdir(directory)
+            os.system("%s siesta < %s | tee %s" % (mpi, inpname, output))
+            os.chdir("../")
 
 
     def analysis(self, directory="tmp-siesta-static", inpname="static.fdf", output="static.out"):
@@ -73,44 +97,54 @@ class static_run:
         plt.show()
         os.chdir("../")
 
-    def converge_cutoff(self, emin, emax, step, directory="tmp-siesta-converge-cutoff"):
-
-        if os.path.exists(directory):
-            shutil.rmtree(directory)
-        os.mkdir(directory)
+    def converge_cutoff(self, emin, emax, step, directory="tmp-siesta-converge-cutoff",
+            mpi="", runopt="gen", electrons={}):
+        if runopt == "gen" or runopt == "genrun":
+            if os.path.exists(directory):
+                shutil.rmtree(directory)
+            os.mkdir(directory)
         
-        for element in self.system.xyz.specie_labels:
-            shutil.copyfile("%s.psf" % element, os.path.join(directory, "%s.psf" % element))
+            for element in self.system.xyz.specie_labels:
+                shutil.copyfile("%s.psf" % element, os.path.join(directory, "%s.psf" % element))
 
-        n_test = int((emax - emin) / step)
-        for i in range(n_test + 1):
-            meshcutoff = int(emin + i * step)
-            self.electrons.params["MeshCutoff"] = meshcutoff
-            with open(os.path.join(directory, "cutoff-%d.fdf" % meshcutoff), 'w') as fout:
-                self.system.to_fdf(fout)
-                self.electrons.to_fdf(fout)
-                self.properties.to_fdf(fout)
-        # run
-        os.chdir(directory)
-        for i in range(n_test + 1):
-            meshcutoff = int(emin + i * step)
-            os.system("siesta < cutoff-%d.fdf | tee cutoff-%d.out" % (meshcutoff, meshcutoff))
-        os.chdir("../")
-        # analysis
-        os.chdir(directory)
-        for i in range(n_test + 1):
-            meshcutoff = int(emin + i * step)
-            out_f_name = "cutoff-%d.out" % meshcutoff
-            os.system("cat %s | grep 'Total =' >> energy-cutoff.data" % out_f_name)
-        cutoff = [emin + i * step for i in range(n_test + 1) ]
-        energy = []
-        with open("energy-cutoff.data", 'r') as fin:
-            for line in fin:
-                energy.append(float(line.split()[3]))
-        import matplotlib.pyplot as plt
-        plt.plot(cutoff, energy)
-        plt.show()
-        os.chdir("../")
+            self.electrons.set_params(electrons)
+            self.electrons.dm["UseSaveDM"] = "false"
 
+            n_test = int((emax - emin) / step)
+            for i in range(n_test + 1):
+                meshcutoff = int(emin + i * step)
+                self.electrons.params["MeshCutoff"] = meshcutoff
+                self.system.label = "siesta-" + str(meshcutoff)
+                with open(os.path.join(directory, "cutoff-%d.fdf" % meshcutoff), 'w') as fout:
+                    self.system.to_fdf(fout)
+                    self.electrons.to_fdf(fout)
+                    self.properties.to_fdf(fout)
+        if runopt == "run" or runopt == "genrun":
+            # run
+            os.chdir(directory)
+            for i in range(n_test + 1):
+                meshcutoff = int(emin + i * step)
+                os.system("%s siesta < cutoff-%d.fdf | tee cutoff-%d.out" % (mpi, meshcutoff, meshcutoff))
+            os.chdir("../")
+            # analysis
+            os.chdir(directory)
+            for i in range(n_test + 1):
+                meshcutoff = int(emin + i * step)
+                out_f_name = "cutoff-%d.out" % meshcutoff
+                os.system("cat %s | grep 'Total =' >> energy-cutoff.data" % out_f_name)
+            cutoff = [emin + i * step for i in range(n_test + 1) ]
+            energy = []
+            with open("energy-cutoff.data", 'r') as fin:
+                for line in fin:
+                    energy.append(float(line.split()[3]))
+            plt.plot(cutoff, energy, marker="o")
+            plt.title("Energy against MeshCutoff")
+            plt.xlabel("MeshCutoff (Ry)")
+            plt.ylabel("Energy (Ry)")
+            plt.show()
+            plt.savefig("energy-cutoff.png")
+            os.chdir("../")
+    
     def set_spin(self, spin="non-polarized"):
         self.electrons.set_spin(spin)
+
